// Features/reports/Types/IPositionForReport.ts
export interface IPositionForReport {
    identifier: string;
    idPosition: number;
    positionName: string;
    minSalary: number;
    maxSalary: number;
    departmentName: string;
    filledPositions: number;
    availablePositions: number;
    numberOfPositions: number;
    idCompany: number;
}